refactor bundle generation code from index.js to internal/bundlegen.js
